# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "BundleFileResponse": {
        "properties": {
          "content": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "sha256": {
            "type": "string"
          },
          "sizeBytes": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "BundleResponse": {
        "properties": {
          "contentHash": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/BundleFileResponse"
            },
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "nginxInstanceId": {
            "format": "uuid",
            "type": "string"
          },
          "sequence": {
            "format": "int64",
            "type": "integer"
          },
          "siteCount": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "BundleSummaryResponse": {
        "properties": {
          "contentHash": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "sequence": {
            "format": "int64",
            "type": "integer"
          },
          "siteCount": {
            "format": "int32",
            "type": "integer"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ConfigurationPreviewResponse": {
        "description": "The configuration a site would produce, and what would change on the host.",
        "properties": {
          "changed": {
            "type": "boolean"
          },
          "changedPaths": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "currentContentHash": {
            "type": "string"
          },
          "proposedContentHash": {
            "type": "string"
          },
          "siteConfiguration": {
            "type": "string"
          }
        },
        "required": [
          "changed",
          "changedPaths",
          "proposedContentHash",
          "siteConfiguration"
        ],
        "type": "object"
      },
...
      "DeploymentEventResponse": {
        "description": "One deployment phase.",
        "properties": {
          "at": {
            "format": "date-time",
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "result": {
            "type": "string"
          }
        },
        "required": [
          "at",
          "phase",
          "result"
        ],
        "type": "object"
      },
      "DeploymentResponse": {
        "description": "One deployment and its phase history.",
        "properties": {
          "attempt": {
            "format": "int32",
            "type": "integer"
          },
          "configBundleId": {
            "format": "uuid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/DeploymentEventResponse"
            },
            "type": "array"
          },
          "finishedAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "nginxInstanceId": {
            "format": "uuid",
            "type": "string"
          },
          "nginxTestOutput": {
            "type": "string"
          },
          "previousBundleId": {
            "format": "uuid",
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "trigger": {
            "type": "string"
          }
        },
        "required": [
          "attempt",
          "createdAt",
          "createdBy",
          "events",
          "id",
          "nginxInstanceId",
          "status",
          "trigger"
        ],
        "type": "object"
      },
...
      "GrantPreviewResponse": {
        "description": "What a grant would reach, before it is created.",
        "properties": {
          "coversFutureDomains": {
            "type": "boolean"
          },
          "matched": {
            "items": {
              "$ref": "#/components/schemas/MatchedSiteResponse"
            },
            "type": "array"
          },
          "newlyVisibleToSubject": {
            "format": "int64",
            "type": "integer"
          },
          "totalMatched": {
            "format": "int64",
            "type": "integer"
          },
          "truncated": {
            "type": "boolean"
          }
        },
        "required": [
          "coversFutureDomains",
          "matched",
          "newlyVisibleToSubject",
          "totalMatched",
          "truncated"
        ],
        "type": "object"
      },
      "GrantRequest": {
        "properties": {
          "domainPattern": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "scopeGroupId": {
            "format": "uuid",
            "type": "string"
          },
          "scopeSiteId": {
            "format": "uuid",
            "type": "string"
          },
          "scopeType": {
            "type": "string"
          },
          "subjectRef": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string"
          },
          "subjectType": {
            "type": "string"
          }
        },
        "required": [
          "level",
          "scopeType",
          "subjectRef",
          "subjectType"
        ],
        "type": "object"
      },
...
      "MatchedSiteResponse": {
        "description": "A site a grant would reach.",
        "properties": {
          "alreadyReachable": {
            "type": "boolean"
          },
          "domain": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "alreadyReachable",
          "domain",
          "id",
          "name"
        ],
        "type": "object"
      },
...
      "PageResponseDeploymentResponse": {
        "properties": {
          "content": {
            "items": {
              "$ref": "#/components/schemas/DeploymentResponse"
            },
            "type": "array"
          },
          "page": {
            "format": "int32",
            "type": "integer"
          },
          "size": {
            "format": "int32",
            "type": "integer"
          },
          "totalElements": {
            "format": "int64",
            "type": "integer"
          },
          "totalPages": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "UpstreamCheckResponse": {
        "properties": {
          "error": {
            "type": "string"
          },
          "reachable": {
            "type": "boolean"
          },
          "target": {
            "type": "string"
          }
        },
        "required": [
          "reachable",
          "target"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer-jwt": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Management API for reverse-proxy sites across one or more NGINX hosts.\n\nAuthentication is OIDC: obtain an access token from Keycloak and send it as\na bearer token. Errors are RFC 9457 problem documents; branch on the stable\n`type` URI rather than on the human-readable `detail`.\n",
    "license": {
      "name": "Proprietary"
    },
    "title": "Easy NGINX Admin API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/config-bundles/{id}": {
      "get": {
        "description": "Private key material is excluded from the file list at every permission level.",
        "operationId": "bundle",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BundleResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Inspect a configuration bundle",
        "tags": [
          "Deployments"
        ]
      }
    },
    "/api/v1/deployments": {
      "get": {
        "operationId": "list_5",
        "parameters": [
          {
            "in": "query",
            "name": "nginxInstanceId",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "items": {
                "type": "string"
              },
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "format": "int32",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 25,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponseDeploymentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List deployments",
        "tags": [
          "Deployments"
        ]
      }
    },
    "/api/v1/deployments/{id}": {
      "get": {
        "operationId": "get_3",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Fetch one deployment with its phase history",
        "tags": [
          "Deployments"
        ]
      }
    },
    "/api/v1/deployments/{id}/rollback": {
      "post": {
        "description": "Queues a new deployment of the previous bundle. Rollback is never automatic: the only automatic revert is the agent restoring the previous release when a reload fails, which is a refusal to leave the host broken.",
        "operationId": "rollback",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Re-activate the configuration this deployment replaced",
        "tags": [
          "Deployments"
        ]
      }
    },
...
    "/api/v1/nginx-instances/{id}/bundles": {
      "get": {
        "operationId": "bundles",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BundleSummaryResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Recent configuration bundles for an instance, newest first",
        "tags": [
          "Deployments"
        ]
      }
    },
    "/api/v1/nginx-instances/{id}/deploy": {
      "post": {
        "operationId": "deployInstance",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Deploy an entire NGINX instance",
        "tags": [
          "Deployments"
        ]
      }
    },
...
    "/api/v1/permissions/preview": {
      "post": {
        "description": "Authorised exactly like creating the grant. Answers the question a wildcard pattern hides: '*.example.com' at MANAGE is one line of configuration and authority over every subdomain, including ones created later. The list is what the scope reaches today; coversFutureDomains says whether it will keep admitting new ones.",
        "operationId": "preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/GrantPreviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Show what a grant would reach, without creating it",
        "tags": [
          "Permissions"
        ]
      }
    },
...
    "/api/v1/proxy-sites/{id}/deploy": {
      "post": {
        "description": "Returns 202: the bundle is rendered from current state and applied by the dispatcher, so the response arrives before the host has been changed. Poll the deployment for the outcome.",
        "operationId": "deploySite",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DeploymentResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Deploy the NGINX instance this site belongs to",
        "tags": [
          "Deployments"
        ]
      }
    },
...
    "/api/v1/proxy-sites/{id}/preview": {
      "get": {
        "description": "Renders without deploying, and reports which files would change on the host.",
        "operationId": "preview_1",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationPreviewResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Show the configuration this site would produce",
        "tags": [
          "Deployments"
        ]
      }
    },
...
    "/api/v1/proxy-sites/{id}/upstream-check": {
      "get": {
        "description": "Advisory. NGINX Open Source resolves upstream names once when the configuration loads, so a reachable upstream now may still fail later; this catches a typo or a service that is not running. The probe runs on the NGINX host, which is where the answer means something and where NGINX will connect from \u2014 the management server never dials a user-supplied address.",
        "operationId": "checkUpstreams",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UpstreamCheckResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Ask the host whether this site's upstreams accept a connection",
        "tags": [
          "Deployments"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearer-jwt": []
    }
  ],
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost:8080"
    }
  ],
...
}